Исправление пары багов

+++ app/database/models/member.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, Integer, BigInteger, VARCHAR
+from sqlalchemy import Column, INTEGER, BIGINT, VARCHAR
 
-from database import DatabaseModel
+from database.models import DatabaseModel
 
 
 class Member(DatabaseModel):
-    id = Column(Integer(), autoincrement=True, primary_key=True)
-    tg_id = Column(BigInteger(), unique=True, nullable=False)
+    id = Column(INTEGER(), autoincrement=True, primary_key=True)
+    tg_id = Column(BIGINT(), unique=True, nullable=False)
     full_name = Column(VARCHAR(254), nullable=False)
     username = Column(VARCHAR(254), nullable=True, server_default=None)
